+++ app/containers/Search/index.js
@@ -22,41 +22,7 @@ import {
 import debounce from 'lodash/debounce';
 
 import { get, post } from '../../core/http/post';
-
-const useStyles = makeStyles(theme => ({
-  root: {
-    marginTop: theme.spacing(8),
-    display: 'flex',
-    flexDirection: 'column',
-    alignItems: 'center',
-  },
-  searchContainer: {
-    padding: '2px 4px',
-    display: 'flex',
-    alignItems: 'center',
-    width: 600,
-  },
-  resultContainer: {
-    alignItems: 'center',
-    width: 600,
-  },
-
-  suggestionItem: {
-    borderBottom: '1px #AAAAAA solid',
-  },
-  input: {
-    marginLeft: 8,
-    flex: 1,
-  },
-  iconButton: {
-    padding: 10,
-  },
-  divider: {
-    width: 1,
-    height: 28,
-    margin: 4,
-  },
-}));
+import useStyles from './styles';
 
 const Search = props => {
   const classes = useStyles();
