+++ app/containers/Search/styles.js
@@ -1,0 +1,38 @@
+import { makeStyles } from '@material-ui/core';
+
+const useStyles = makeStyles(theme => ({
+  root: {
+    marginTop: theme.spacing(8),
+    display: 'flex',
+    flexDirection: 'column',
+    alignItems: 'center',
+  },
+  searchContainer: {
+    padding: '2px 4px',
+    display: 'flex',
+    alignItems: 'center',
+    width: 600,
+  },
+  resultContainer: {
+    alignItems: 'center',
+    width: 600,
+  },
+
+  suggestionItem: {
+    borderBottom: '1px #AAAAAA solid',
+  },
+  input: {
+    marginLeft: 8,
+    flex: 1,
+  },
+  iconButton: {
+    padding: 10,
+  },
+  divider: {
+    width: 1,
+    height: 28,
+    margin: 4,
+  },
+}));
+
+export default useStyles;
